requirements.txt
numpy==1.17.3
Pillow==6.2.0
PySide2==5.13.1
shiboken2==5.13.1
